chore(vitest): move vue config into vitest project
