+++ .github/workflows/module.yml
@@ -52,9 +52,6 @@ jobs:
       - name: Test
         run: pnpm run test run
 
-      - name: Test (vue)
-        run: pnpm run test:vue run
-
       - name: Build
         run: pnpm run build
